css/css-fonts/format-specifiers-variations.html

<!DOCTYPE HTML>
<html>
<head>
<title>CSS Test: Supported format specifiers should load</title>
<link rel="help" href="https://drafts.csswg.org/css-fonts-4/#src-desc"/>
<meta name="assert" content="Supported format specifiers should load"/>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
setup({
    explicit_done: true
});

var formatSpecifiers = [
    "woff",
    "truetype",
    "opentype",
    "woff2",
    "woff-variations",
    "truetype-variations",
    "opentype-variations",
    "woff2-variations"
];

var keywordFormatSpecifiers = [
    "opentype",
    "truetype",
    "woff",
    "woff2"
];

var defaultAhemSrc = 'url("/fonts/Ahem.ttf") format(INSERT_FORMAT)';

function runTestOnFormatSpecifiers(formats, keyword, expectFail) {
    for (var i = 0; i < formats.length; ++i) {
        promise_test((testDetails) => {
            var familyName = "load_ahem_" + i;
            var format = testDetails.properties.format
            if (!keyword) {
                format = '"' + format + '"';
            }
            var ahemSrcFormat = defaultAhemSrc.replace("INSERT_FORMAT", format);
            var fontFace = new FontFace(familyName, ahemSrcFormat);
            if (!expectFail) {
                return fontFace.load();
            } else {
                return promise_rejects_dom(testDetails, "NetworkError", fontFace.load());
            }
        }, (expectFail ? "Do not load" : "Load") + " Ahem with" + (keyword ? " keyword" : "") + " format " + formats[i], {
            "format": formats[i]
        });
    }
}

runTestOnFormatSpecifiers(formatSpecifiers, false, false);
runTestOnFormatSpecifiers(formatSpecifiers.map(x => "xyz" + x), false, true);
runTestOnFormatSpecifiers(formatSpecifiers.map(x => x + "xyz"), false, true);
runTestOnFormatSpecifiers(formatSpecifiers.map(x => x.slice(0, -2)), false, true);
runTestOnFormatSpecifiers(formatSpecifiers.map(x => x.slice(2)), false, true);
runTestOnFormatSpecifiers(formatSpecifiers.map(x => x.slice(0, x.length / 3) + x.slice(x.length / 3 + 1)), false, true);

runTestOnFormatSpecifiers(keywordFormatSpecifiers, true, false);

done();
</script>
</body>
</html>
